// frontend/src/api/models/ContentResponse.ts
export interface ContentResponse {
    /**
     * 
     * @type {string}
     * @memberof ContentResponse
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof ContentResponse
     */
    createdBy?: string;
    /**
     * 
     * @type {Date}
     * @memberof ContentResponse
     */
    createdDate?: Date;
    /**
     * 
     * @type {string}
     * @memberof ContentResponse
     */
    lastModifiedBy?: string;
    /**
     * 
     * @type {Date}
     * @memberof ContentResponse
     */
    lastModifiedDate?: Date;
    /**
     * 
     * @type {string}
     * @memberof ContentResponse
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof ContentResponse
     */
    slug?: string;
    /**
     * 
     * @type {string}
     * @memberof ContentResponse
     */
    body?: string;
    /**
     * 
     * @type {string}
     * @memberof ContentResponse
     */
    excerpt?: string;
    /**
     * 
     * @type {string}
     * @memberof ContentResponse
     */
    coverImageUrl?: string;
    /**
     * 
     * @type {string}
     * @memberof ContentResponse
     */
    status?: ContentResponseStatusEnum;
    /**
     * 
     * @type {Date}
     * @memberof ContentResponse
     */
    publishedAt?: Date;
    /**
     * 
     * @type {Set<string>}
     * @memberof ContentResponse
     */
    tags?: Set<string>;
    /**
     * 
     * @type {StaffResponse}
     * @memberof ContentResponse
     */
    author?: StaffResponse;
}


/**
 * @export
 */
export const ContentResponseStatusEnum = {
    Draft: 'DRAFT',
    Published: 'PUBLISHED',
    Archived: 'ARCHIVED'
} as const;
export type ContentResponseStatusEnum = typeof ContentResponseStatusEnum[keyof typeof ContentResponseStatusEnum];